api: Simplify a few callers' names now that there's no collision.
See b8010deca for discussion of names like these which have been
contorted to avoid collisions with names from the API bindings.

We already took care of some of the more egregious cases of this in
b8010deca and subsequent commits, when we originally introduced the
`import api` style and applied it in a few places.  But there are a
few left; now that (in the parent commit) we've systematically
applied the still-better `import * as api` style everywhere,
take care of those stragglers too.