RealmScreen: Log a warning on failure to connect to server.
This can be perfectly normal... but it can also be a symptom of
configuration trouble or a bug somewhere, and in that case the details
will be helpful for debugging.  And it's extremely low-volume.

In particular on Android this causes the details to be shown in
`adb logcat`, and in `adb logcat -s ReactNativeJS:W` for a reasonably
low-noise view.  That's helpful for debugging an install of the app
in the field.

The use of `console.warn` is exactly what we want here, though the
juxtaposition with our logging functions in `utils/logging` is a bit
messy.  Really we should rationalize those into a coherent set of
logging functions at the various severities, and then we'd use one of
those here.  Meanwhile, just silence the lint warning.