logging: Let `console.log` and friends do their work in release mode, too.
The effect of these is, on Android, to log the messages to the device
log with tag `ReactNativeJS`.  So this command finds them:
  adb logcat -s ReactNativeJS:V
I'm not sure if it does anything helpful on iOS, though I'd hope so.

With this change, everything we send to Sentry via `logErrorRemotely`,
plus a few other things where we call `console` methods directly, is
now available in the device log.  Often the most readily available
repro of an issue is by one of us on our normal devices, or by a user
who happens to be comfortable with `adb`; and in that case this is a
much, much easier way to see the logs (and iterate on the logging!)
than trying to locate the corresponding report on Sentry, if there
even is one.

The output is remarkably low-noise.  At startup I get one I message,
and one W from the new community NetInfo package we switched to in
b909c0a22.  Then so far the only output I've found a way to provoke is

 (a) on failing to connect to a server at RealmScreen; and

 (b) one previously-unknown bug!  It's in the `lastNarrow` logic;
     fix coming next.

This transform to disable logging was added a couple of years ago in
861dcfba7.  There's no discussion there, and no explanation; so I
don't know why someone thought it was a good idea.  Perhaps they
didn't realize that the output can be found with `adb logcat`, and
thought it was only available in the debug console, in debug builds?
In any case, it certainly didn't lead to us doing a good job of having
effective logging thereafter to be able to debug issues.