Revert "android build: Use our patched react-native-webview to cut noise in log."
This reverts commit 1982f3fa8183502b5388dec0377c518fdeedf8a9.

`react-native-webview` relies on a `lib/` directory which is not
automatically built, but instead usually downloaded from `npm`.
Since the source repo doesn't contain `lib/`, this breaks the build.

As this is presently only cosmetic, reverting until either we figure
out how to build it as needed, or (per the previous commit) update
`react-native`.