api: Adjust Reaction type to describe complexity in API.
When messages are fetched from server via GET `/messages`, then it
contains `id` as key in `user` of the reaction object. While reaction
add event has `user_id` in it.

We haven't been using the `user_id` property anywhere, which is why
we never noticed at runtime that it was missing.