NavButton: Cut long-disused `unreadCount` feature.
The last use of this, I believe, was cut in 58b6bb768 -- two years
ago, in the era of the right and left drawer.

Trying it out recently, it doesn't look right: the count is shown way
off to the upper right, some ways away from the icon.  Cut it out to
avoid confusion.