ComponentWithOverlay: Move unrelated functionality to callers.
The touch functionality has nothing to do with the overlay.  Both
this component and its callers are simpler to understand if it's
done separately by the callers.

One of the two callers turns out not even to be using it!  A
`Touchable` with no props set turns out (from looking at the
implementation) to work out to just a `View`... so simplify it to
that.  (The `View` might in turn be redundant, but that requires
some care to confirm.)