search: refactor search screen to be more principled
Perform search queries at query-receipt time in SearchMessagesScreen,
rather than in its child component SearchMessagesCard as a side-effect
of rendering the latter.

This is a brute-force hoisting with no (intentional) behavioral
changes, to be refined in the following series of commits.

(Note that the removal of `throttle` is not a behavioral change: it
doesn't actually do anything the way it's currently being called.)