isMessageInNarrow: Fix check for starred or mentioned messages.
Newly arriving messages did not appear in the is:starred narrow if
they were starred, or in the is:mentioned narrow if they contained an
@-mention (either direct or wildcard).

In isMessageInNarrow (src/utils/narrow.js), instead of checking if
message.type is 'starred' or 'mentioned' -- which didn't work, because
it never has those values -- we now correctly check for the relevant
values in message.flags .  Also add tests for this logic.

We also assert (with an error) that message.flags is defined.  It
should be; see parent commit, and if not, this will help us catch the
bug in dev.  Add a test for this exception, and update existing tests
to include message.flags.

Fixes: #3592