deps: Cut lodash.throttle.
It turns out the one place we were using this library was in
search... where, as 1c51173f5 observed, we were misusing it so that
it didn't actually do anything!

So, cut it from package.json and our libdefs.  It survives in
yarn.lock as an indirect dependency.

Also cut reference to it in a comment, written for a version of this
code that included throttling.