android notif [nfc]: Factor out tryGetReactInstanceManager logic.
This lets us encapsulate the knowledge of this RN quirk, and our
workaround, a bit more cleanly than by having an early-return
condition and a comment.

Also, hooray for Kotlin extension methods.  If we had to write this
as a function with an argument, that'd be OK... but this framing is
more appropriate.