api: Add new ApiError type and constructor-helper
Add a type to be used for API errors. Replace existing `makeApiError`
helper with the slightly-more-rigorous `makeErrorFromApi`.

In theory, one could make `ApiError` generic over `T: ApiResponseErrorData`,
and refine errors to their appropriate types at runtime. However, this
would be a lot of machinery (both type-level and runtime) for effectively
no present benefit: we don't have a use yet for the additional data on any
of the precise types.

Based loosely on a series of commits by Boris Yankov in PR #3484.