deps/jest: Add `jsdom-global` to dev dependencies
For a later test, we'll want to be able to set up the test environment
to appear as though it's a browser at a specified URL. Jest currently
uses `jsdom` as its mock-browser environment, but doesn't expose the
`jsdom`-provided facilities needed to set the mock `location`.

Fortunately, Jest does allow alternate mock environments. We install
one which is identical to Jest's default environment, except that it
does expose that capability.

(Note that there _are_ dependency-free methods to tell Jest to tell
JSDOM what URL(s) to use; however, they involve a hierarchy of
configuration files and spooky action-at-a-distance, rather than an
obvious function call in the relevant script.)