gradle: Add (inactive) `shrinkResources` option for release builds
The Android Gradle plugin comes with a binary-assets shrinker.  We
don't use many binary assets, so at the moment it has little effect
other than wasting cycles during build time; but it costs almost
nothing to make note of the option for later.

As of this commit, on my development box, the resulting release .apk
size is:

   default:             22,416,749 bytes
   Proguard only:       21,099,887 bytes
   Proguard + shrink:   21,020,091 bytes