ensure-coreutils: Make suitable to use in normal dev scripts.
A need for `fmt` may be fairly niche, but `readlink -f` is awfully
basic and widely useful in shell scripting.  Fortunately, with the
insight that on Windows none of this is an issue, we can press ahead
more widely with a clear conscience.

Mostly this expands the comments; also adds a `brew install` mention
to the error/help message.

Also make the test a bit pickier, checking `readlink` too.

A draft version prefixed `fmt` and `readlink` with `command`, to avoid
seeing aliases or shell functions... but this apparently triggered a
bug in the Bash 3.2.57 on a Mac:
  $ bash -ec 'command false || echo f; echo done'
    # no output
Ah well.

Suggested-by: Ray Kraesig <rkraesig@zulipchat.com>