docs: Small further edits to WSL instructions.
The most interesting of these (though none are all that interesting)
are to deduplicate with discussion elsewhere.

In particular, versions of things like Node risk going stale.
I suspect the JDK version mentioned may already be stale; didn't
want to just change it without testing, but at least it's now
flagged as a thing to watch.