rewriteImageUrls: Use URL.origin rather than ponyfill.
It turns out that in fact this is already present in Chrome 37!
The compatibility table on MDN was more cautious than the truth.
I've sent a PR upstream to fix it:
  https://github.com/mdn/browser-compat-data/pull/5078

Discovered this because I went to rewrite the ponyfill to directly
reflect the spec:
  https://url.spec.whatwg.org/#dom-url-origin
I considered narrowing the test at the top to look for `undefined`;
so to double-check that that's the value `.origin` gave in Chrome
versions that didn't have the feature, I fired up the app in an
emulated Android 6 device, with Chrome 44... and the value wasn't
`undefined` because in fact the feature was present there after all.

Well, that simplifies things a bit.

In the process of testing the other features the table had said
arrived in Chrome 52 -- all of which also turned out to date at least
to Chrome 37 -- I also discovered a new bug in the existing ponyfill:

  > new URL("https://me:pw0rd@example.org/foo").origin
  "https://example.org"

The ponyfill would have given "https://me:pw0rd@example.org",
including the username and password.  Unlikely this would have ever
mattered to this particular code, though.