Emoji [nfc]: Fold ImageEmoji helper into main component.
This is a very small amount of code, so this makes the parent
component barely longer.  And the most interesting information the
helper expressed -- namely, the width and height of the image --
are actually an important part of the interface, because they
correspond to the size passed to `UnicodeEmoji` in the other case.
So that information wasn't actually encapsulated anyway, and should
be made explicit in the parent component one way or another.