AccountDetails [nfc]: Cut redundant `void` for state type.
When the component doesn't have any state -- as most of our components
don't -- specifying a state type is just clutter.

Moreover `void` is actually the default type for this parameter,
so saying `void` here has no effect at all.  Take it out.

(This is the only place we do this.)