webview build [nfc]: Cut speculative scaffolding comments.
We might in the future have complex things we're doing here, which
may need a complex structure which may have elements in common with
the one suggested by these comments.

But at present we're doing something quite simple -- just copying a
few files.  It isn't productive to try to decide on a complex future
structure before we know what concrete needs we'll have for it.
Until then, this empty scaffolding just gets in the way of reading
what we're actually doing, and of extending it in the directions
called for by our current known needs.