vscode: Remove obsolete config setting for Prettier extension.
We use prettier-eslint for formatting, and we have a setting in our
VS Code config to tell that to the Prettier extension we use.  A new
version of the extension (3.8) now detects our prettier-eslint setup,
making this option redundant... and as long the option is still there
it spams a warning box on every save.  Remove the relevant (or
rather, now-irrelevant) setting from our workspace config.

Upstream doc:
  https://github.com/prettier/prettier-vscode/blob/62004a02b/README.md#prettier-linter-integration-advanced

Developers with an older version of the extension will need to update
it to keep getting the right formatting behavior on save.  By default
VS Code auto-updates extensions, in which case just reloading the
editor may be enough.

[greg: revised commit message]