notif: Resend push tokens after logout and login.
In general we avoid constantly telling the server our push token
(the token provided by APNs or FCM for this app-device pair) again
and again, by storing in `ackedPushToken` the token the server has
already heard.

But since zulip/zulip@6407d0b1f, if the user resets their API key:
  https://zulipchat.com/api/api-keys#invalidate-an-api-key
and so logs out all their devices, the server will also forget the
push tokens (or tell the bouncer to do so) so that the notification
behavior matches.

That means if we get logged out and then the user logs back in again,
we may need to send the token anew.  So, do that.

Fixes: #3695