sentry: Revert upgrade, returning to v0.37 from v1.x.
This reverts commit ec101ddcb "deps: upgrade Sentry".

It turns out Sentry's API changed more than we realized across this
upgrade, so we'll need to do more work than this to complete the
upgrade properly.  As is, we believe the Sentry integration in this
version of the upgrade doesn't really work.  Some further discussion
at #3683.

See also commits f0326e756 and 86cdcbfe3, which reverted other steps
of the Sentry upgrade series.

On the other hand commits ab5d6a5aa and b4c11e02a, though they were
also part of that series, are independently helpful refactors that
don't rely on the upgrade, so those changes remain in place.