sentry: add custom Flow types
Internally, the Sentry JavaScript SDK uses TypeScript. Fortunately,
the types used are not terribly advanced; in almost all cases they can
just be copied over. Porting them to Flow is more of an exercise in
pointer-chasing than anything else.

The port is not complete, but does cover all existing uses and a few
hypothetical ones.