sentry (0.x): remove all use
Despite sharing many of the same method names, the APIs for the React
Native Sentry SDK in versions 0.x and 1.x are almost completely
incompatible, in ways both subtle and gross. It's probably best not to
think of a transition from 0.x to 1.x as an upgrade, but rather as a
move to a completely different framework.

In preparation for this move, eliminate all calls into Sentry 0.x from
the codebase. Their functionality will be restored in an upcoming
commit, after its replacement with Sentry 1.x.