docs/style: On mentioning fixed and related issues.
I've written the key points here a few times ad hoc, including:
  https://github.com/zulip/zulip-mobile/pull/3496#issuecomment-495414420
  https://github.com/zulip/zulip-mobile/pull/3554#issuecomment-544596492
(plus some useful Q&A in subsequent comments on the former), and
others I don't have on hand.

So, write it up a bit more fully in a place that's easy to find
and to point people at.

Also describe the `Fixes: #1234` syntax.  We've often also said
`Fixes #1234.`; but I've shifted to preferring the other form as
we've started writing kernel-style `Frobbed-by:` lines, because
it fits so naturally as one more of those.