eslint: enable leading underscores to denote "private" fields
The TC39 class fields proposal [1] -- stage 3, but already implemented
and available in Chrome 74 -- allows for truly private member
variables. This feature has also been supported by Babel since at
least version 7.3.

Alas, the intersection of private fields and *Flow* is a much more
poorly-supported area. There are several known issues:

* prettier/prettier#5288
* babel/babel-eslint#688
* facebook/flow#7877

A workaround exists for the first one (Babel is now usable as the
parser), but not the latter two.

As an approximation, then, we allow leading underscores in property
names until this feature is available.

[1]: https://github.com/tc39/proposal-class-fields