docs/build-run: Replace `yarn reinstall` with explicit commands.
This has always been a pretty ad-hoc remedy that sometimes works.
The details of it have a better home written explicitly in the
troubleshooting entry than wrapped up into an opaque, and
misleadingly polished-looking, custom command.

On top of that, the command as written in package.json doesn't work
on the Windows Command Prompt.  That... shouldn't actually matter!
Our .npmrc instructs Yarn to use Bash (which Windows users have
thanks to Git) -- see 6c25beeb0.  But sometimes that seems to go
wrong; and that's a problem for a command which is specifically
documented to use in a debugging session where things are already
going confusingly wrong.