flow: add JSONable type
Add (a reasonable approximation of) `JSONable`, to be used in the
following commits.

Mostly, we've been passing around `mixed` where a JSONable value is
expected, and expecting `JSON.stringify()` to always turn it into a
`string`. Unfortunately, this isn't what it does -- in particular,
`JSON.stringify(undefined) === undefined`.

Flow 0.98.0 fixes the incorrect definition of `JSON.stringify` that
we'd been relying on. [0] We'll need to take extra steps to ensure
that `undefined` doesn't leak.

The `JSONable` type introduced here actually goes a step further,
however. When we convert things to JSON, we're generally also
concerned with being able to extract them again later. As such, we
exclude many types from `JSONable` that `JSON.stringify` will happily
accept -- only values that `JSON.parse` will also produce are
recognized as being `JSONable`.

[0] https://github.com/facebook/flow/commit/1f6bcf88c