utils: add "well-typed" "ponyfill" for Object.entries
The ES2017 function `Object.entries` is available in Flow, but its
definition in lib/core.js leaves much to be desired. Add a utility
ponyfill for Object.entries, with a much more convenient type.

This ponyfill is formally unsound; but most users are unlikely to
notice, and at any rate this can be remedied in a future version of
Flow. (See internal comments for more detail.)