ComposeMenu.js: Unsuppress and handle Flow errors with ImagePicker.
Remove $FlowFixMe above import of ImagePicker; change to
`import * as ImagePicker` so ImagePicker is no longer treated as
`any`.

This was hiding an error caused by the `response` param in
this.handleImagePickerResponse not being $ReadOnly, so, make it
$ReadOnly so no inconsistencies can be introduced by modifying it.

Also revealed was the need to type `fileName` to string | void to
match the Response type in react-native-image-picker; do.