api types: Annotate registerForEvents: returns Promise<InitialData>.
At this function's call site, in doInitialFetch, Flow was getting
`any` for the return value, and doing no type-checking on how we
used it, since an exported value without an annotation is treated as
`any`.

We do have a type for what we expect the /register request to return
(i.e., InitialData in 'src/api/initialDataTypes.js'), but we just
weren't using it. InitialData doesn't follow the same naming
convention as, e.g., ApiResponseServerSettings in
getServerSettings.js, but InitialData is OK because it's more
concise.