MessageList: Switch to new React Context API, remove miscStyles.webview.
Per #1946, we're moving to the new React Context API
(https://reactjs.org/docs/context.html).

We’re also isolating the theme colors in their own context type, to
be accessed from there instead of the previously used
`context.styles` object. So, components that consume the theme
colors can and should use `static contextType = ThemeContext;` and,
e.g., this.context.backgroundColor, instead of
this.context.styles.backgroundColor, as was done before.

This change also eliminates the need for miscStyles.webview, which
was included in this.context.styles that MessageList was using, so,
remove that.