unreadNotice: Shrink height to 0 when compose box expands.
The height of UnreadNotice which was stucked on the top was
equal to the `paddingVertical` applied to it. So when it comes to shrink
RN was persisting padding.

Actually, this `View` wrapper is just a add on, not required. Styles can
be directly applied to `Animated.View`. So remove whis extra wrapper
from UnreadNotice, which fixes the issue.

Fixes: #3857