emoji: Cut misleading ignored reference to an emoji font.
Per the docs of the react-native-vector-icons package, this argument
is supposed to be the name of a font.  But there's no sign of a font
by this name actually appearing in the package, so it's hard to see
what passing it could actually mean -- instead, it just misleads the
reader by suggesting that this font is involved somehow.

And, empirically, the behavior is the same when this argument is
removed.  Tested in the reaction-list screen, in an emulated Pixel 3
running Android 10 and in an emulated iPhone 8 running iOS 13.