narrow [nfc]: Rename some surprisingly-narrow predicates to be more explicit.
There are very few places where we really want to distinguish between
PM narrows that are 1:1 and that are with groups.  The more natural
category, both in our implementation logic and in the user experience,
is the one that includes both types of narrows as "PM narrows".

So, instead of having a long complicated name (isPrivateOrGroupNarrow)
for the predicate that tests that category and short ones
(isPrivateNarrow, isGroupNarrow) for those that test the more specific
categories, we should have longer and more explicit names for the
latter and a simpler name for the former.

Here, do the first part of that change:
  isPrivateNarrow -> is1to1PmNarrow
  isGroupNarrow -> isGroupPmNarrow

We'll do the remaining rename in the next commit:
  isPrivateOrGroupNarrow -> isPmNarrow

Done automatically, with:

  $ perl -i -0pe '
        s/isPrivateNarrow/is1to1PmNarrow/g;
        s/isGroupNarrow/isGroupPmNarrow/g
      ' src/**/*.js