ZulipVersion: Add, for server feature checking and Sentry logging.
E.g., a mobile feature could be enabled if the server is running
version 2.1.1 or later, and disabled otherwise.
ZulipVersion.prototype.isAtLeast(otherZulipVersion) should be used
for this.

Also add logic for reporting the major, minor, and patch versions in
Sentry to inform event aggregation.
ZulipVersion.prototype.loggingArray() should be used for this.
.raw() should also be used to send the raw version string to Sentry.