jest: Add Lolex for proper time mocking.
At present, it's not possible (or at least not easy) to test any code
which relies on `Date.now()` for timing purposes. Timer functions such
as `setInterval()` are mockable with Jest's existing mock-timer
functionality, but its mock timers do not include a mock time. [0]

Fortunately, there is work underway to have Jest v25^H^H26 use the
Lolex time-mocking library for its fake timer implementation, which
will do exactly that! Unfortunately, we need that capability now,
rather than in the indefinite future, so we'll jump the gun a bit.

To ease conversion of tests once Lolex is available natively, we use a
small shim taken from the code on Jest's current master branch which
implements Jest's timer-control functions (not projected to change)
in terms of Lolex.

Tests for this shim -- which may also serve, in part, as examples for
Zulip contributors who would like to use Lolex -- are included.

(Additionally, inform Jest that files present in src/__tests__/aux are
not supposed to contain executable tests.)

[0] S2E1, 1967-09-15.

GitHub-PR: https://github.com/zulip/zulip-mobile/pull/3886